Fix issue that "Use touch tone keypad" doesn't work
When selecting "Use touch tone keypad" during call, it expects to start
InCallActivity with dialpad. So will set INTENT_EXTRA_SHOW_DIALPAD as
true in intent extra data. But in "onNewIntent@InCallActivity",
"recreate()" is executed. It will save the dialpad status(it is
invisible currently) to icicle bundle in "onSaveInstanceState". Which
overwrites the dialpad value of intent extra. So dialpad is not shown.

To resolve this issue, use the value of intent extra and icicle bundle
to check whether show dialpad or not.

Test: manual - Checked that the dialpad is displayed when returned to
the call screen by selecting "Use touch tone keypad"
Bug: 64278190

Change-Id: I1a84b52b36ec2f98990874bb37669489bcf117c8
Signed-off-by: AKabhishek <abhishekkaushik.star@gmail.com>